requirements.txt
opencv-python
numpy
face_recognition
faiss-cpu
cryptography
gradio
